Skips metamodel visualization tests if bokeh requirement is not satisfied
